Move FuzzyMatchContainer to a separate class
