+++ src/main/kotlin/com/mituuz/fuzzier/FuzzierVCS.kt
@@ -5,6 +5,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectFileIndex
 import com.intellij.openapi.vcs.changes.ChangeListManager
 import com.mituuz.fuzzier.components.FuzzyFinderComponent
+import com.mituuz.fuzzier.entities.FuzzyMatchContainer
 import org.apache.commons.lang3.StringUtils
 import javax.swing.DefaultListModel
 import javax.swing.SwingUtilities
@@ -24,7 +25,7 @@ class FuzzierVCS : Fuzzier() {
         currentTask?.takeIf { !it.isDone }?.cancel(true)
         currentTask = ApplicationManager.getApplication().executeOnPooledThread {
             component.fileList.setPaintBusy(true)
-            val listModel = DefaultListModel<StringEvaluator.FuzzyMatchContainer>()
+            val listModel = DefaultListModel<FuzzyMatchContainer>()
             val projectFileIndex = ProjectFileIndex.getInstance(project)
             val changeListManager = ChangeListManager.getInstance(project)
             val projectBasePath = project.basePath
