+++ src/main/kotlin/com/mituuz/fuzzier/components/FuzzyComponent.kt
@@ -2,7 +2,7 @@ package com.mituuz.fuzzier.components
 
 import com.intellij.ui.EditorTextField
 import com.intellij.ui.components.JBList
-import com.mituuz.fuzzier.StringEvaluator.FuzzyMatchContainer
+import com.mituuz.fuzzier.entities.FuzzyMatchContainer
 import javax.swing.JPanel
 
 open class FuzzyComponent : JPanel() {
