+++ src/main/kotlin/com/mituuz/fuzzier/components/TestBenchComponent.kt
@@ -13,7 +13,7 @@ import com.intellij.ui.components.JBTextArea
 import com.intellij.uiDesigner.core.GridConstraints
 import com.intellij.uiDesigner.core.GridLayoutManager
 import com.mituuz.fuzzier.StringEvaluator
-import com.mituuz.fuzzier.StringEvaluator.FuzzyMatchContainer
+import com.mituuz.fuzzier.entities.FuzzyMatchContainer
 import org.apache.commons.lang3.StringUtils
 import java.awt.Dimension
 import java.util.*
